# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: JOE Hub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /modules/{id}:
    get:
      operationId: getModule
      tags: [modules]
      summary: Get a module
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Module
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Module"
        "404":
          description: Not found
    patch:
      operationId: updateModule
      tags: [modules]
      summary: Update a module (admin/mentor)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModuleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Module"
        "404":
          description: Not found
    delete:
      operationId: deleteModule
      tags: [modules]
      summary: Delete a module (admin/mentor)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /modules/{id}/lessons:
    get:
      operationId: listModuleLessons
      tags: [lessons]
      summary: List lessons of a module
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of lessons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lesson"
    post:
      operationId: createLesson
      tags: [lessons]
      summary: Create a lesson in a module (admin/mentor)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"

  # ---------- Lessons ----------
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a single notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
        "404":
          description: Not found
...
components:
  schemas:
...
    Module:
      type: object
      required: [id, courseId, title, order, lessonCount, createdAt]
      properties:
        id: { type: integer }
        courseId: { type: integer }
        title: { type: string }
        order: { type: integer }
        lessonCount: { type: integer }
        createdAt: { type: string, format: date-time }
...
    ModuleUpdate:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        order: { type: integer }
    Lesson:
      type: object
      required: [id, moduleId, title, order, createdAt, completed, bookmarked]
      properties:
        id: { type: integer }
        moduleId: { type: integer }
        title: { type: string }
        content: { type: ["string", "null"] }
        videoUrl: { type: ["string", "null"] }
        order: { type: integer }
        completed: { type: boolean }
        bookmarked: { type: boolean }
        createdAt: { type: string, format: date-time }
    LessonInput:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1 }
        content: { type: string }
        videoUrl: { type: string }
        order: { type: integer, default: 0 }
...
    Notification:
      type: object
      required: [id, userId, type, title, body, isRead, createdAt]
      properties:
        id: { type: integer }
        userId: { type: integer }
        type: { type: string }
        title: { type: string }
        body: { type: string }
        link: { type: ["string", "null"] }
        isRead: { type: boolean }
        createdAt: { type: string, format: date-time }
...